requirements.txt
crewai==0.28.8
PyGithub==2.1.1
requests==2.31.0
python-dotenv==1.0.0
pytest==7.4.0